fix: private router 로 변경, form으로 compoment 분리, nickname fetch api 추가
